requirements.txt
Django==5.2
djangorestframework==3.16.0
psycopg2==2.9.10
drf-yasg==1.21.10
